Separate out no-op Unix mount
